Add wallet features #5
- Move staking and transaction history from dashboard to wallet.
- Migrate wallet logic into a WalletService.
- Fix some branding from City Chain to City Hub.
