+++ src/app/components/about/about.component.html
@@ -1,6 +1,7 @@
 <div class="content not-selectable">
     <mat-card class="about-card">
-        <img mat-card-image src="/assets/city/city-chain-gold.png" alt="City Chain">
+        <!-- <img mat-card-image src="" class="about-logo" alt="City Chain"> -->
+        <img src="/assets/city-hub/city-hub-dark.png" class="about-logo" alt="City Hub">
         <mat-card-header>
             <mat-card-title>Version 0.0.1
                 <br>Network: City Chain (Main)</mat-card-title>
