+++ src/app/components/about/about.component.scss
@@ -1,20 +1,31 @@
 .about {
     .about-card {
-        max-width: 700px;
+        max-width: 600px;
         margin-bottom: 30px;
         padding-top: 20px;
       }
     
     .about-card .mat-card-image:first-child
-        {
+    {
         margin-top: -8px;
-        }
+        margin-left: 0 !important;
+        margin-right: 0 !important;
+
+        
+
+    }
     
      .about-nav-title {
         flex: 1 1 auto;
         text-overflow: ellipsis;
         overflow: hidden;
         white-space: nowrap;
+    }
+
+    .about-logo {
+        margin-left: 10px;
+        margin-right: 10px;
+        width: calc(100% - 20px);
     }
     
     .about-nav-caret {
