+++ src/app/components/dashboard/dashboard.component.scss
@@ -9,6 +9,10 @@
         }
     }
 
+    .dashboard-coinmap {
+        min-height: 200px;
+    }
+
     .transaction-details-button
     {
         height: 10px;
@@ -28,6 +32,6 @@
     }
 
     mat-card {
-        margin-bottom: 10px;
+        margin-bottom: 20px;
     }
 }
