+++ src/app/components/dashboard/dashboard.component.ts
@@ -8,6 +8,7 @@ import { Router } from '@angular/router';
 import { WalletInfo } from '../../classes/wallet-info';
 import { DetailsService } from '../../services/details.service';
 import { ApplicationStateService } from '../../services/application-state.service';
+import { WalletService } from '../../services/wallet.service';
 
 
 export interface PeriodicElement {
@@ -39,24 +40,24 @@ const ELEMENT_DATA: PeriodicElement[] = [
 export class DashboardComponent implements OnInit, OnDestroy {
     @HostBinding('class.dashboard') hostClass = true;
 
-    public walletName: string;
-    public coinUnit: string;
-    public confirmedBalance: number;
-    public unconfirmedBalance: number;
-    public transactionArray: TransactionInfo[];
-    private stakingForm: FormGroup;
-    private walletBalanceSubscription: Subscription;
-    private walletHistorySubscription: Subscription;
-    private stakingInfoSubscription: Subscription;
-    public stakingEnabled: boolean;
-    public stakingActive: boolean;
-    public stakingWeight: number;
-    public netStakingWeight: number;
-    public expectedTime: number;
-    public dateTime: string;
-    public isStarting: boolean;
-    public isStopping: boolean;
-    public hasBalance = false;
+    // public walletName: string;
+    // public coinUnit: string;
+    // public confirmedBalance: number;
+    // public unconfirmedBalance: number;
+    // public transactionArray: TransactionInfo[];
+    // private stakingForm: FormGroup;
+    // private walletBalanceSubscription: Subscription;
+    // private walletHistorySubscription: Subscription;
+    // private stakingInfoSubscription: Subscription;
+    // public stakingEnabled: boolean;
+    // public stakingActive: boolean;
+    // public stakingWeight: number;
+    // public netStakingWeight: number;
+    // public expectedTime: number;
+    // public dateTime: string;
+    // public isStarting: boolean;
+    // public isStopping: boolean;
+    // public hasBalance = false;
 
     displayedColumns: string[] = ['position', 'name', 'weight'];
     dataSource = ELEMENT_DATA;
@@ -66,286 +67,39 @@ export class DashboardComponent implements OnInit, OnDestroy {
         private router: Router,
         public appState: ApplicationStateService,
         private detailsService: DetailsService,
+        public wallet: WalletService,
         private fb: FormBuilder) {
-        this.buildStakingForm();
+
+        // Make sure we update wallet at higher frequency.
+        this.wallet.active = true;
     }
 
     ngOnInit() {
-        this.startSubscriptions();
-        this.walletName = this.globalService.getWalletName();
-        this.coinUnit = this.globalService.getCoinUnit();
+        //this.startSubscriptions();
+        // this.walletName = this.globalService.getWalletName();
+        // this.coinUnit = this.globalService.getCoinUnit();
     }
 
     ngOnDestroy() {
-        this.cancelSubscriptions();
+        // When navigate away from dashboard, we'll make the wallet update more slowly.
+        this.wallet.active = false;
+        //this.cancelSubscriptions();
     }
 
-    private buildStakingForm(): void {
-        this.stakingForm = this.fb.group({
-            "walletPassword": ["", Validators.required]
-        });
-    }
 
-    public openTransactionDetails(transaction: TransactionInfo) {
 
-        this.detailsService.show(transaction);
+    // private cancelSubscriptions() {
+    //     if (this.walletBalanceSubscription) {
+    //         this.walletBalanceSubscription.unsubscribe();
+    //     }
 
-        //const modalRef = this.modalService.open(TransactionDetailsComponent, { backdrop: "static", keyboard: false });
-        //modalRef.componentInstance.transaction = transaction;
-    }
+    //     if (this.walletHistorySubscription) {
+    //         this.walletHistorySubscription.unsubscribe();
+    //     }
 
-    private cancelSubscriptions() {
-        if (this.walletBalanceSubscription) {
-            this.walletBalanceSubscription.unsubscribe();
-        }
+    //     if (this.stakingInfoSubscription) {
+    //         this.stakingInfoSubscription.unsubscribe();
+    //     }
+    // }
 
-        if (this.walletHistorySubscription) {
-            this.walletHistorySubscription.unsubscribe();
-        }
-
-        if (this.stakingInfoSubscription) {
-            this.stakingInfoSubscription.unsubscribe();
-        }
-    }
-
-    private startSubscriptions() {
-        this.getWalletBalance();
-        this.getHistory();
-        this.getStakingInfo();
-    }
-
-    private getWalletBalance() {
-        let walletInfo = new WalletInfo(this.globalService.getWalletName());
-        this.walletBalanceSubscription = this.apiService.getWalletBalance(walletInfo)
-            .subscribe(
-                response => {
-                    if (response.status >= 200 && response.status < 400) {
-                        let balanceResponse = response.json();
-                        //TO DO - add account feature instead of using first entry in array
-                        this.confirmedBalance = balanceResponse.balances[0].amountConfirmed;
-                        this.unconfirmedBalance = balanceResponse.balances[0].amountUnconfirmed;
-                        if ((this.confirmedBalance + this.unconfirmedBalance) > 0) {
-                            this.hasBalance = true;
-                        } else {
-                            this.hasBalance = false;
-                        }
-                    }
-                },
-                error => {
-                    console.log(error);
-                    if (error.status === 0) {
-                        this.cancelSubscriptions();
-                        // this.genericModalService.openModal(null, null);
-                    } else if (error.status >= 400) {
-                        if (!error.json().errors[0]) {
-                            console.log(error);
-                        } else {
-                            if (error.json().errors[0].description) {
-                                // this.genericModalService.openModal(null, error.json().errors[0].message);
-                            } else {
-                                this.cancelSubscriptions();
-                                this.startSubscriptions();
-                            }
-                        }
-                    }
-                }
-            )
-            ;
-    };
-
-    // todo: add history in seperate service to make it reusable
-    private getHistory() {
-        let walletInfo = new WalletInfo(this.globalService.getWalletName());
-        let historyResponse;
-        this.walletHistorySubscription = this.apiService.getWalletHistory(walletInfo)
-            .subscribe(
-                response => {
-                    if (response.status >= 200 && response.status < 400) {
-                        //TO DO - add account feature instead of using first entry in array
-                        if (!!response.json().history && response.json().history[0].transactionsHistory.length > 0) {
-                            historyResponse = response.json().history[0].transactionsHistory;
-                            this.getTransactionInfo(historyResponse);
-                        }
-                    }
-                },
-                error => {
-                    console.log(error);
-                    if (error.status === 0) {
-                        this.cancelSubscriptions();
-                        // this.genericModalService.openModal(null, null);
-                    } else if (error.status >= 400) {
-                        if (!error.json().errors[0]) {
-                            console.log(error);
-                        }
-                        else {
-                            if (error.json().errors[0].description) {
-                                // this.genericModalService.openModal(null, error.json().errors[0].message);
-                            } else {
-                                this.cancelSubscriptions();
-                                this.startSubscriptions();
-                            }
-                        }
-                    }
-                }
-            )
-            ;
-    };
-
-    private getTransactionInfo(transactions: any) {
-        this.transactionArray = [];
-
-        for (let transaction of transactions) {
-            let transactionType;
-            if (transaction.type === "send") {
-                transactionType = "sent";
-            } else if (transaction.type === "received") {
-                transactionType = "received";
-            } else if (transaction.type === "staked") {
-                transactionType = "staked";
-            }
-            let transactionId = transaction.id;
-            let transactionAmount = transaction.amount;
-            let transactionFee;
-            if (transaction.fee) {
-                transactionFee = transaction.fee;
-            } else {
-                transactionFee = 0;
-            }
-            let transactionConfirmedInBlock = transaction.confirmedInBlock;
-            let transactionTimestamp = transaction.timestamp;
-            let transactionConfirmed;
-
-            this.transactionArray.push(new TransactionInfo(transactionType, transactionId, transactionAmount, transactionFee, transactionConfirmedInBlock, transactionTimestamp));
-        }
-    }
-
-    private startStaking() {
-        this.isStarting = true;
-        this.isStopping = false;
-        let walletData = {
-            name: this.globalService.getWalletName(),
-            password: this.stakingForm.get('walletPassword').value
-        }
-        this.apiService.startStaking(walletData)
-            .subscribe(
-                response => {
-                    if (response.status >= 200 && response.status < 400) {
-                        this.stakingEnabled = true;
-                        this.stakingForm.patchValue({ walletPassword: "" });
-                    }
-                },
-                error => {
-                    this.isStarting = false;
-                    this.stakingEnabled = false;
-                    this.stakingForm.patchValue({ walletPassword: "" });
-                    if (error.status === 0) {
-                        // this.genericModalService.openModal(null, null);
-                    } else if (error.status >= 400) {
-                        if (!error.json().errors[0]) {
-                            console.log(error);
-                        } else {
-                            // this.genericModalService.openModal(null, error.json().errors[0].message);
-                        }
-                    }
-                }
-            )
-            ;
-    }
-
-    private stopStaking() {
-        this.isStopping = true;
-        this.isStarting = false;
-        this.apiService.stopStaking()
-            .subscribe(
-                response => {
-                    if (response.status >= 200 && response.status < 400) {
-                        this.stakingEnabled = false;
-                    }
-                },
-                error => {
-                    if (error.status === 0) {
-                        // this.genericModalService.openModal(null, null);
-                    } else if (error.status >= 400) {
-                        if (!error.json().errors[0]) {
-                            console.log(error);
-                        } else {
-                            // this.genericModalService.openModal(null, error.json().errors[0].message);
-                        }
-                    }
-                }
-            )
-            ;
-    }
-
-    private getStakingInfo() {
-        this.stakingInfoSubscription = this.apiService.getStakingInfo()
-            .subscribe(
-                response => {
-                    if (response.status >= 200 && response.status < 400) {
-                        const stakingResponse = response.json();
-                        this.stakingEnabled = stakingResponse.enabled;
-                        this.stakingActive = stakingResponse.staking;
-                        this.stakingWeight = stakingResponse.weight;
-                        this.netStakingWeight = stakingResponse.netStakeWeight;
-                        this.expectedTime = stakingResponse.expectedTime;
-                        this.dateTime = this.secondsToString(this.expectedTime);
-                        if (this.stakingActive) {
-                            this.isStarting = false;
-                        } else {
-                            this.isStopping = false;
-                        }
-                    }
-                },
-                error => {
-                    if (error.status === 0) {
-                        // this.genericModalService.openModal(null, null);
-                    } else if (error.status >= 400) {
-                        if (!error.json().errors[0]) {
-                            console.log(error);
-                        } else {
-                            // this.genericModalService.openModal(null, error.json().errors[0].message);
-                        }
-                    }
-                }
-            )
-            ;
-    }
-
-    private secondsToString(seconds: number) {
-        let numDays = Math.floor(seconds / 86400);
-        let numHours = Math.floor((seconds % 86400) / 3600);
-        let numMinutes = Math.floor(((seconds % 86400) % 3600) / 60);
-        let numSeconds = ((seconds % 86400) % 3600) % 60;
-        let dateString = "";
-
-        if (numDays > 0) {
-            if (numDays > 1) {
-                dateString += numDays + " days ";
-            } else {
-                dateString += numDays + " day ";
-            }
-        }
-
-        if (numHours > 0) {
-            if (numHours > 1) {
-                dateString += numHours + " hours ";
-            } else {
-                dateString += numHours + " hour ";
-            }
-        }
-
-        if (numMinutes > 0) {
-            if (numMinutes > 1) {
-                dateString += numMinutes + " minutes ";
-            } else {
-                dateString += numMinutes + " minute ";
-            }
-        }
-
-        if (dateString === "") {
-            dateString = "Unknown";
-        }
-
-        return dateString;
-    }
 }
