+++ src/app/components/load/load.component.html
@@ -1,7 +1,7 @@
 <div class="splash-container not-selectable">
 
     <div class="">
-        <img src="/assets/{{appState.coin}}/logo-full-dark.png" class="load-logo">
+        <img src="/assets/city-hub/city-hub-dark.png" class="load-logo">
     </div>
 
     <ng-template [ngIf]="!loading" [ngIfElse]="spinner">
