+++ src/app/components/merchants/merchants.component.html
@@ -1,0 +1,1 @@
+<iframe src="https://coinmap.org/" class="merchant-map"></iframe>
