+++ src/app/components/merchants/merchants.component.scss
@@ -1,0 +1,8 @@
+.merchant {
+    
+    .merchant-map {
+        width: 100%;
+        height: 100%;
+        border: 0;
+    }
+}
