+++ src/app/components/wallet/wallet.module.ts
@@ -7,6 +7,8 @@ import { ReceiveModule } from './receive/receive.module';
 import { SendModule } from './send/send.module';
 import { WalletRoutingModule } from './wallet-routing.module';
 import { PaymentModule } from './payment/payment.module';
+import { ReactiveFormsModule, FormsModule } from '@angular/forms';
+import { AppSharedModule } from '../../shared/app-shared.module';
 
 @NgModule({
     imports: [
@@ -14,6 +16,9 @@ import { PaymentModule } from './payment/payment.module';
         RouterModule,
         ReceiveModule,
         SendModule,
+        ReactiveFormsModule,
+        FormsModule,
+        AppSharedModule,
         PaymentModule,
         MaterialModule,
         WalletRoutingModule
