+++ src/styles/styles.scss
@@ -77,6 +77,10 @@ body {
 	overflow: hidden;
 }
 
+.content-container-full {
+	height: 100% !important;
+}
+
 .content
 {
 	margin: 30px;
